Change base Docker image to 3.6-adoptopenjdk-8
The previous Docker base image is no longer maintained. The maven images
are now based on images from adoptopenjdk project. This image is based
on the same jdk-8 as the previous one and on the same maven-3.6 branch
as the previous one. This image is maintained by upstream and updated
regularly.

maven:3.6-adoptopenjdk-8 is based on Ubuntu 18.04 LTS version and
therefore the needed packages and their install commands had to be
updated but the installed packages now offer the necessary hcidump and
hcitool for RuuviCollector to run.